Use file instead of link in dependencies
If you use link in dependencies of package.json, you will get an error when you run yarn publish.
Use file instead.

There are situations where you want it to behave as a link rather than a file when running locally.
In that case, instead of writing link in package.json, use yarn link to create a symbolic link.