Introduce VRT test
There is a test method called Visual Regression Testing.
It is a test method that takes a screenshot of the actual drawing result in the browser and compares it to detect regressions.

There is a possibility that it will not work well if there are differences in the browser between the local environment and the test environment.